third_party/blink/web_tests/virtual/plz-dedicated-worker-disabled/external/wpt/html/cross-origin-embedder-policy/credentialless/cache-storage.https.window_dedicated_worker-expected.txt

This is a testharness.js-based test.
[PASS] [dedicated_worker] unsafe-none => unsafe-none
[FAIL] [dedicated_worker] unsafe-none => credentialless
  assert_equals: expected "error" but got "retrieved"
[PASS] [dedicated_worker] unsafe-none => credentialless (omit)
[PASS] [dedicated_worker] unsafe-none => credentialless + CORP
[FAIL] [dedicated_worker] unsafe-none => require-corp
  assert_equals: expected "error" but got "retrieved"
[FAIL] [dedicated_worker] unsafe-none => require-corp (omit)
  assert_equals: expected "error" but got "retrieved"
[PASS] [dedicated_worker] unsafe-none => require-corp + CORP
[PASS] [dedicated_worker] credentialless => unsafe-none
[PASS] [dedicated_worker] credentialless => credentialless
[FAIL] [dedicated_worker] credentialless => require-corp
  assert_equals: expected "error" but got "retrieved"
[PASS] [dedicated_worker] credentialless => require-corp + CORP
[PASS] [dedicated_worker] require_corp => unsafe-none
[PASS] [dedicated_worker] require_corp => credentialless
[PASS] [dedicated_worker] require_corp => require-corp
Harness: the test ran to completion.
